Refactor and address PR review#3 issues
