+++ nrpy/equations/general_relativity/tests/InitialData_Cartesian_Kasner.py
@@ -1,0 +1,32 @@
+from mpmath import mpf  # type: ignore
+
+trusted_dict = {
+    "BU_0": mpf("0.0"),
+    "BU_1": mpf("0.0"),
+    "BU_2": mpf("0.0"),
+    "KDD_0": mpf("-0.824366945161734356298277469121573"),
+    "KDD_1": mpf("0.0"),
+    "KDD_2": mpf("0.0"),
+    "KDD_3": mpf("0.0"),
+    "KDD_4": mpf("-0.0547671704401983067783504658653579"),
+    "KDD_5": mpf("0.0"),
+    "KDD_6": mpf("0.0"),
+    "KDD_7": mpf("0.0"),
+    "KDD_8": mpf("-0.304150815002603243262540520225332"),
+    "alpha": mpf("1.0"),
+    "betaU_0": mpf("0.0"),
+    "betaU_1": mpf("0.0"),
+    "betaU_2": mpf("0.0"),
+    "gammaDD_0": mpf("0.797505322933511254768416295005559"),
+    "gammaDD_1": mpf("0.0"),
+    "gammaDD_2": mpf("0.0"),
+    "gammaDD_3": mpf("0.0"),
+    "gammaDD_4": mpf("0.987938765744869697912937436811132"),
+    "gammaDD_5": mpf("0.0"),
+    "gammaDD_6": mpf("0.0"),
+    "gammaDD_7": mpf("0.0"),
+    "gammaDD_8": mpf("0.930984919936852995575498206582433"),
+    "x": mpf("0.481822907157808399958298650744837"),
+    "y": mpf("0.106164922885331747259840540209552"),
+    "z": mpf("0.680168862988865807750471503823064"),
+}
