+++ nrpy/examples/exact_kasner_evolution.py
@@ -128,11 +128,25 @@ if num_fisheye_transitions is not None:
     for parname, value in fisheye_param_defaults.items():
         par.adjust_CodeParam_default(parname, value)
 
+kasner_param_guard = r"""
+const REAL kasner_sum_p = commondata->KASNER_p1 + commondata->KASNER_p2 + commondata->KASNER_p3;
+const REAL kasner_sum_p2 = commondata->KASNER_p1 * commondata->KASNER_p1
+                         + commondata->KASNER_p2 * commondata->KASNER_p2
+                         + commondata->KASNER_p3 * commondata->KASNER_p3;
+const REAL kasner_abs_sum_err = (kasner_sum_p > 1.0) ? (kasner_sum_p - 1.0) : (1.0 - kasner_sum_p);
+const REAL kasner_abs_sum2_err = (kasner_sum_p2 > 1.0) ? (kasner_sum_p2 - 1.0) : (1.0 - kasner_sum_p2);
+if (kasner_abs_sum_err > 1.0e-14 || kasner_abs_sum2_err > 1.0e-14) {
+  fprintf(stderr, "Error: KASNER_p1, KASNER_p2, KASNER_p3 must satisfy p1+p2+p3=1 and p1^2+p2^2+p3^2=1.\\n");
+  exit(1);
+}
+"""
+
 BHaH.general_relativity.initial_data.register_CFunction_initial_data(
     IDtype=IDtype,
     IDCoordSystem=IDCoordSystem,
     set_of_CoordSystems=set_of_CoordSystems,
     ID_persist_struct_str="",
+    populate_ID_persist_struct_str=kasner_param_guard,
 )
 
 BHaH.numerical_grids_and_timestep.register_CFunctions(
